+++ server/src/settings.ts
@@ -1,3 +1,4 @@
+import * as _ from 'lodash';
 import * as os from 'os';
 import { DiagnosticSeverity } from 'vscode-languageserver/node';
 
@@ -22,6 +23,27 @@ export function propertiesPlatform() {
 }
 
 export type SeverityLevel = DiagnosticSeverity | string;
+
+export namespace VS_DiagnosticSeverity {
+    export function from(value: SeverityLevel): DiagnosticSeverity {
+        if (_.isNumber(value)) {
+            return value as DiagnosticSeverity;
+        }
+
+        if (!_.isString(value)) {
+            throw TypeError(`The diagnostic code ${value} was neither a number nor string!`);
+        }
+
+        switch (value) {
+            case 'Error': return DiagnosticSeverity.Error;
+            case 'Warning': return DiagnosticSeverity.Warning;
+            case 'Information': return DiagnosticSeverity.Information;
+            case 'Hint': return DiagnosticSeverity.Hint;
+            default:
+                throw RangeError(`The diagnostic code ${value} has no mapping to DiagnosticSeverty.`);
+        }
+    }
+}
 
 export interface CppCheckSeverityMaps {
     error: SeverityLevel;